zmeny dle pozadavku

+++ common/src/main/java/cz/mycom/veeam/portal/model/Subtenant.java
@@ -24,10 +24,10 @@ public class Subtenant {
     private String uid;
 
     @Column
-    private Long quota;
+    private int quota;
 
     @Column
-    private Long usedQuota;
+    private int usedQuota;
 
     @Column
     private boolean enabled;
